backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.8.0
python-dotenv>=1.0.0
firebase-admin>=6.4.0
dedalus-labs>=0.1.0a9